requirements.txt
streamlit>=1.30
openai>=1.3.5
faiss-cpu==1.7.4
sentence-transformers==2.2.2
yfinance==0.2.36
pandas==1.5.3
plotly==5.19.0
transformers==4.30.2
tokenizers==0.13.3
huggingface_hub==0.15.1
